Create Construction-Department
In our cooperation, you are able to Explore the demands of the construction industry in the 21st century and understand how to address those needs. Discover the roles and opportunities available to you in the construction industry, and the pathways to get you into the sector. Explore research methods for construction management and understand construction project cost management. Explore the critical issues faced by the construction industry today and discover sustainable approaches to resolving them.
Team Prospects:
•  To develop learners’ understanding of sustainable construction innovation and improvement as a collaborative, inter-organizational learning process – And to develop their capabilities to lead others to this same understanding. 
•  To inspire participants to “think big”, by employing this type of collaborative learning. 
•  To cultivate participants’ sense of agency, by examining ways that diverse teams of leaders collaborate in designing and leading construction and infrastructure sys